# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subscription API",
        "version": "1.0"
    },
    "paths": {
        "/api/subscribers": {
            "get": {
                "summary": "Get all subscribers",
                "operationId": "7a4ec4a0bed9cb1bb0f252aa5fde1ac4",
                "responses": {
                    "200": {
                        "description": "A list of subscribers"
                    }
                }
            },
            "post": {
                "summary": "Create a new subscriber",
                "operationId": "cdbdfb57ab947416e0ecf36c82ed29e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscriber"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscriber created"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/subscribers/{id}": {
            "get": {
                "summary": "Get a specific subscriber",
                "operationId": "cb2ebc56c624405a54492302dbb13aee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscriber found"
                    },
                    "404": {
                        "description": "Subscriber not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific subscriber",
                "operationId": "f7edc93245402315d3f961da3ec023a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscriber"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscriber updated"
                    },
                    "404": {
                        "description": "Subscriber not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific subscriber",
                "operationId": "d5182899aaacc15b1b2a04f7808dd746",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subscriber deleted"
                    }
                }
            }
        },
        "/api/subscriptions": {
            "get": {
                "summary": "Get all subscriptions",
                "operationId": "67df8b16e2d36e7aab97ce19a17f9355",
                "responses": {
                    "200": {
                        "description": "A list of subscriptions"
                    }
                }
            },
            "post": {
                "summary": "Create a new subscription",
                "operationId": "a71648bdc4f0b03d53530ab22adf00cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscription"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/subscriptions/{id}": {
            "get": {
                "summary": "Get a specific subscription",
                "operationId": "6b9cf0d3971878009b6d4f0a5598dae9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription found"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific subscription",
                "operationId": "bac8be53d2f4233f44fd60abb494230b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscription"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific subscription",
                "operationId": "06c2e9a3211bd2ad09833e2f9b53af8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subscription deleted"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Subscriber": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "example@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
            "Subscription": {
                "properties": {
                    "service": {
                        "type": "string",
                        "example": "Email Service"
                    },
                    "topic": {
                        "type": "string",
                        "example": "Newsletter"
                    },
                    "payload": {
                        "type": "object",
                        "example": {
                            "key": "value"
                        }
                    },
                    "expired_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-13T13:31:23.069Z"
                    },
                    "subscriber_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        }
    }
}